+++ bake.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 # Released under the MIT License.
-# Copyright, 2024, by Samuel Williams.
+# Copyright, 2024-2026, by Samuel Williams.
 
 # Update the project documentation with the new version number.
 #
@@ -9,4 +9,11 @@
 def after_gem_release_version_increment(version)
 	context["releases:update"].call(version)
 	context["utopia:project:update"].call
+end
+
+# Create a GitHub release for the given tag.
+#
+# @parameter tag [String] The tag to create a release for.
+def after_gem_release(tag:, **options)
+	context["releases:github:release"].call(tag)
 end