+++ examples/benchmark/falcon.rb
@@ -2,7 +2,7 @@
 # frozen_string_literal: true
 
 # Released under the MIT License.
-# Copyright, 2019-2025, by Samuel Williams.
+# Copyright, 2019-2026, by Samuel Williams.
 
 require "falcon/environment/rack"
 require "falcon/environment/self_signed_tls"