+++ examples/grpc/server.rb
@@ -1,6 +1,9 @@
 #!/usr/bin/env ruby
 # frozen_string_literal: true
 
+# Released under the MIT License.
+# Copyright, 2025-2026, by Samuel Williams.
+
 require "grpc"
 require_relative "my_service_services_pb"